sharing worker

+++ src/AsyncTask.js
@@ -1,44 +1,49 @@
-var inherits          = require( 'util' ) .inherits,
-    EventEmitter      = require( 'events' ).EventEmitter,
-    BackgroundWorker  = require( 'background-worker' )
-
+var BackgroundWorker = require( '../../background-worker' )
+var uuid             = require( 'uuid' )
 
 module.exports = AsyncTask
 
 /*
  * @class AsyncTask
- * @extends EventEmitter
  * @author Jørn Andre Tangen @gorillatron
 */
 function AsyncTask( options ) {
-  EventEmitter.apply( this, arguments )
-
   options = typeof options != 'undefined' ? options : {}
+
+  this.__uuid = uuid.v4()
+
   this.__hasExecuted = false
   this.__keepAlive = options.keepAlive
-  this.__boundArguments = []
+  this.__sharingworker = false
 
   this.doInBackground = options.doInBackground ? options.doInBackground : null
   this.importScripts = options.importScripts ? options.importScripts : []
 
+  if( options.worker ) {
+    this.__sharingworker = true
+    this.setWorker( options.worker )
+  }
 }
-
-inherits( AsyncTask, EventEmitter )
 
 /*
  * Setup a background-worker/BackgroundWorker
  * @public
  * @function
 */
-AsyncTask.prototype.setupWorker = function() {
-  if( !this.worker ) {
-    this.worker = new BackgroundWorker({})
-    this.worker.importScripts = this.importScripts
-    this.worker.define( 'doInBackground', this.doInBackground.toString() )
-    this.worker.start()
-  }
-  return this.worker
+AsyncTask.prototype.setWorker = function( worker ) {
+  this._worker = worker
+  this._worker.importScripts = this.importScripts
+  this._worker.define( this.__uuid + '::doInBackground', this.doInBackground.toString() )
+  return this._worker
 }
+
+/*
+* Array slice
+* @private
+* @function
+* @returns Array
+*/
+var slice = Array.prototype.slice
 
 /*
  * Execute the background job on a worker
@@ -53,16 +58,19 @@ AsyncTask.prototype.execute = function() {
     throw new Error( 'Cannot execute a allready executed AsyncTask' )
   }
 
+  if( !this._worker ) {
+    this.setWorker( new BackgroundWorker({}) )
+  }
+
   this.__hasExecuted = true
-  this.emit( 'execute' )
-  this.setupWorker()
 
-  worker = this.worker
-  args = Array.prototype.slice.call( arguments )
-  taskPromise = this.worker.run( 'doInBackground', args )
+  worker = this._worker
+  args = slice.call( arguments )
+  taskPromise = worker.run( this.__uuid + '::doInBackground', args )
 
-  if( !this.__keepAlive )
+  if( !this.__keepAlive && !this.__sharingworker ) {
     taskPromise.finally(function() { worker.terminate() })
+  }
 
   return taskPromise
 }
